+++ scripts/run_cdk_example.sh
@@ -30,7 +30,7 @@ code_uri="s3://$ARGO_CLOUDOPS_BUILD_BUCKET/cdk-typescript-example.tar.gz"
 execute_container_image_uri='argocloudops/argo-cloudops-cdk:1.99.0'
 workflow_name=argo-cloudops-single-step-vault-aws
 
-$CLI sync \
+$CLI workflow --type sync \
     -a execute='--no-color --require-approval never' \
     -e AWS_REGION="us-west-2",CODE_URI="$code_uri",VAULT_ADDR="$VAULT_ADDR" \
     -f cdk \