+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 
+## Unreleased
+* Adds operations API for git based sync/diff
+* Fix typos in CLI command description
+
 ## v0.0.2 (2021-05-06)
-+ Set service port via environment variable, default 8443
+* Set service port via environment variable, default 8443
 
 ## v0.0.1 (2021-05-06)
-+ Initial release
+* Initial release